+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Todas as atualizações a partir de 30/05/206 devem observar os principios [Mant
 ## 0.3.0-dev - 
 
 ### Added
-- Nothing
+- Geração local do DANFSe v2.0 nacional conforme NT 008, com QR Code, blocos de identificação, tributação, totalizadores, informações complementares, marcas d'água de cancelamento/substituição e canhoto opcional.
 
 ### Deprecated
 - Nothing